Обработка запроса на получение данных

+++ VirtualLegoRobot/Assets/Scripts/Data/Display.cs
@@ -4,13 +4,13 @@ namespace Assets.Scripts.Data
     public class Display
     {
         public DisplayTypes DisplayType;
-        public double X1, X2;
-        public double Y1, Y2;
+        public float X1, X2;
+        public float Y1, Y2;
         public bool ClearScreen;
         public string Text;
         public bool Invert;
         public int Size;
-        public double Radius;
+        public float Radius;
         public bool Fill;
 
     }
