+++ VirtualLegoRobot/Assets/Scripts/Data/Distance.cs
@@ -3,9 +3,9 @@ namespace Assets.Scripts.Data
 {
     public class Distance
     {
-        public double Value { get; private set; }
+        public float Value { get; private set; }
         public UnitTypes Unit { get; set; }
 
-        public Distance(double value, UnitTypes unit) { Value = value; Unit = unit; }
+        public Distance(float value, UnitTypes unit) { Value = value; Unit = unit; }
     }
 }
