+++ VirtualLegoRobot/Assets/Scripts/Data/Motor.cs
@@ -5,9 +5,9 @@ namespace Assets.Scripts.Data
     public class Motor
     {
         public char Port { get; private set; }
-        public double Power { get; private set; }
+        public float Power { get; private set; }
 
-        public Motor(char port, double power)
+        public Motor(char port, float power)
         {
             Port = port;
             Power = power;
